+++ database/seeders/AppSettingSeeder.php
@@ -66,5 +66,7 @@ class AppSettingSeeder extends Seeder
                 $setting
             );
         }
+
+        AppSetting::clearCache();
     }
 }