+++ resources/views/partials/lang/_init.blade.php
@@ -28,6 +28,7 @@
         try {
             localStorage.setItem("data-kt-lang", locale);
             document.cookie = "kt_lang=" + locale + ";path=/;max-age=31536000;SameSite=Lax";
+            document.cookie = "data-kt-lang=" + locale + ";path=/;max-age=31536000;SameSite=Lax";
         } catch (e) {}
 
         window.KTLanguageConfig = {